% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demand}
\alias{demand}
\title{Childcare demand locations}
\format{
A data frame with 4 columns:
\describe{
\item{id}{Sector identifier.}
\item{x}{X coordinate.}
\item{y}{Y coordinate.}
\item{demand}{Demand value.}
}
}
\usage{
demand
}
\description{
A data frame containing demand locations and the number of children.
}
\keyword{internal}
