% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Compute a distance matrix}
\usage{
distance_matrix(
  supply,
  demand,
  supply_x = "x",
  supply_y = "y",
  demand_x = "x",
  demand_y = "y",
  min_distance = 400
)
}
\arguments{
\item{supply}{A data frame containing supply locations.}

\item{demand}{A data frame containing demand locations.}

\item{supply_x}{Name of the x-coordinate column in \code{supply}.}

\item{supply_y}{Name of the y-coordinate column in \code{supply}.}

\item{demand_x}{Name of the x-coordinate column in \code{demand}.}

\item{demand_y}{Name of the y-coordinate column in \code{demand}.}

\item{min_distance}{Minimum distance value. Distances below this value are replaced.}
}
\value{
A numeric matrix with supply units in rows and demand units in columns.
}
\description{
Compute a distance matrix
}
