% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratio_map}
\alias{ratio_map}
\title{Sector-level childcare ratio map}
\format{
An sf object with sector geometries and ratio variables.
}
\usage{
ratio_map
}
\description{
An sf object containing childcare supply, demand and local ratio by sector.
}
\keyword{datasets}
