% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supply}
\alias{supply}
\title{Childcare supply locations}
\format{
A data frame with 4 columns:
\describe{
\item{id}{Childcare facility identifier.}
\item{x}{X coordinate.}
\item{y}{Y coordinate.}
\item{supply}{Supply value.}
}
}
\usage{
supply
}
\description{
A data frame containing childcare locations and capacity.
}
\keyword{internal}
